Single Script/requirements_new.txt
astroid==2.5.6
cycler==0.10.0
isort==5.8.0
joblib==0.17.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
matplotlib==3.4.2
mccabe==0.6.1
numpy==1.19.4
pandas==1.2.4
Pillow==8.2.0
pylint==2.8.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.1
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.3
wrapt==1.12.1
xgboost==1.2.1
